Подключает webpack и разделяет проект на модули
