+++ .gitignore
@@ -1,4 +1,4 @@
-# Файлы и папки операционной системы
+﻿# Файлы и папки операционной системы
 .DS_Store
 Thumbs.db
 
@@ -13,3 +13,7 @@ node_modules/
 
 # Папка с собранными файлами проекта
 build/
+
+# Файлы сборки
+public/bundle.js
+public/bundle.js.map
